flask_app/requirements.txt
flask
paho-mqtt
influxdb-client
dotenv